stacks
# Stacks Operated Exploration

## Goals
Develop and explore models that generate organized complexity using stacks as a data structure.

## Methodology
We'll start with a simple stack and look at how stacking and unstacking operations can produce emerging patterns over iterations. How these operations can contribute to organized complexity will be evaluated.

## Model Rule
The model consists of a series of stacking and unstacking operations that, based on certain parameter evolve to generate an emerging pattern.

## Expected Results
We look for patterns of complexity that are not explicitly programmed, but rather emerge from the interaction of stacking operations.